Moved ModelInfo to toplevel class
